USI hardware support
When the `Display::Sda` is equal to `att85::sda` and `Display::Scl` is
equal to `att::scl`, the USI(Universal Serial Interface) hardware is
used to optimize the `detail::send_byte()` function. The program will
be smaller and faster than a version that only uses a sofware solution
to send a byte.